Criando service para tratar integracao e passando a chama-lo no controller
