+++ src/Controllers/IntegratorController.php
@@ -1,11 +1,9 @@
 <?php
 namespace SonnyBlaine\Integrator\Controllers;
 
-use SonnyBlaine\Integrator\Services\RequestService;
-use SonnyBlaine\Integrator\Services\SourceService;
+use SonnyBlaine\Integrator\Services\IntegratorService;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
-use OldSound\RabbitMqBundle\RabbitMq\Producer as RabbitProducer;
 
 /**
  * Class IntegratorController
@@ -14,34 +12,17 @@ use OldSound\RabbitMqBundle\RabbitMq\Producer as RabbitProducer;
 class IntegratorController
 {
     /**
-     * @var SourceService Service class for Source
+     * @var IntegratorService
      */
-    protected $sourceService;
-
-    /**
-     * @var RequestService Service class for Request
-     */
-    protected $requestService;
-
-    /**
-     * @var RabbitProducer
-     */
-    protected $rabbitProducer;
+    protected $integratorService;
 
     /**
      * IntegratorController constructor.
-     * @param SourceService $sourceService
-     * @param RequestService $requestService
-     * @param RabbitProducer $rabbitProducer
+     * @param IntegratorService $integratorService
      */
-    public function __construct(
-        SourceService $sourceService,
-        RequestService $requestService,
-        RabbitProducer $rabbitProducer
-    ) {
-        $this->sourceService = $sourceService;
-        $this->requestService = $requestService;
-        $this->rabbitProducer = $rabbitProducer;
+    public function __construct(IntegratorService $integratorService)
+    {
+        $this->integratorService = $integratorService;
     }
 
     /**
@@ -53,13 +34,7 @@ class IntegratorController
     public function integrateAction(Request $request, string $sourceIdentifier, string $queryParameter)
     {
         try {
-            $source = $this->sourceService->findByIdentifier($sourceIdentifier);
-
-            $sourceRequest = $this->requestService->createSourceRequest($source, $queryParameter);
-
-            $this->requestService->createDestinationRequest($sourceRequest);
-
-            $this->rabbitProducer->publish($sourceRequest->getId());
+            $sourceRequest = $this->integratorService->integrate($sourceIdentifier, $queryParameter);
 
             return new JsonResponse([
                 'requestId' => $sourceRequest->getId(),
